Move schema related tests to pinot-spi
